+++ dataprovider/foxsports/scraper/eventdata/scraper_mlb_probable_starting_pitcher_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/scraper/matchup"
-	"github.com/lightning-dabbler/sportscrape/util/runner/eventdata"
-	matchuputil "github.com/lightning-dabbler/sportscrape/util/runner/matchup"
+	"github.com/lightning-dabbler/sportscrape/runner/eventdata"
+	mr "github.com/lightning-dabbler/sportscrape/runner/matchup"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -24,9 +24,9 @@ func TestMLBProbableStartingPitcher(t *testing.T) {
 		matchup.ScraperSegmenter(&foxsports.GeneralSegmenter{Date: "2024-10-25"}),
 	)
 
-	matchuprunner := matchuputil.NewRunner(
-		matchuputil.RunnerName("MLB Matchups"),
-		matchuputil.RunnerScraper(matchupScraper),
+	matchuprunner := mr.NewRunner(
+		mr.RunnerName("MLB Matchups"),
+		mr.RunnerScraper(matchupScraper),
 	)
 
 	matchups := matchuprunner.RunMatchupsScraper()
